% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assay.window.R
\docType{data}
\name{assay.window}
\alias{assay.window}
\title{Data to calculate an assay window}
\format{A data frame with 40 rows and 3 variables:
\describe{
  \item{A:}{Values for the first outcome variable.}
  \item{B:}{Values for the second outcome variable.}
  \item{condition:}{Indicates if a sample is a positive or
negative control.}  }}
\description{
Simulated data for two variables to calculate an assay
window between positive and negative control samples.
}

