% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CV.R
\docType{data}
\name{CV}
\alias{CV}
\title{Data to estimate the coefficient of variation}
\format{A data frame with 15 rows and 3 variables:
\describe{
  \item{Sample:}{Sample identification number.}
  \item{A:}{The first outcome variable.}
  \item{B:}{The second outcome variable.}
}}
\description{
Simulated data for two outcome variables.
}
\details{
The data contains five samples with three measurements
made on each sample; these are "technical replicates" and do not
contribute to the sample size. Which of the two outcome variables,
A or B, would make a better primary outcome?
}

