% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/locomotor.R
\docType{data}
\name{locomotor}
\alias{locomotor}
\title{Locomotor activity rat data}
\format{A data frame with 282 rows and 4 variables:
\describe{
  \item{drug:}{Drug or control indicator.}
  \item{animal:}{Rat identification number.}
  \item{time:}{Time point of observation, in increments of 15 minutes.}
  \item{dist:}{Distance travelled.}  }}
\description{
Distance travelled by rats in the open field test for
a drug and control condition.
}
\details{
Forty-seven rats were randomised to a control or drug
condition. Rats were injected with either a drug or saline at time
= 0. Locomotor activity in the open field was recorded for 90
minutes. The distance travelled in each 15 minute interval is
reported.
}

